Refactor CLI args schema help
