+++ AGENTS.md
@@ -12,8 +12,10 @@
   pushed commit needs a fix, create a new forward commit and push normally,
   unless the user explicitly instructs otherwise.
 - **Never** use abbreviations for variable names, aside from common
-  abbreviations (e.g., `i` for loop variables), acronyms (e.g., `url`), and
-  idiomatic library namespace aliases (e.g., `v` for Valibot).
+  abbreviations (e.g., `i` for loop variables, `pkg` for package metadata),
+  acronyms (e.g., `url`), and idiomatic or project-standard namespace/import
+  aliases. Allowed aliases include `v` for Valibot and `cs` for vanilla-extract
+  style namespace imports from `.css.ts` files.
 - **Always use braces** for `if`/`else`/`for`/`while` blocks, even single-line
   bodies.
